requirements.txt
grpcio-tools
grpcio
magika